reduce db code from vm
